pyproject.toml
[project]
name = "ms-reid"
version = "0.1.0"
description = "Graduation project for person re-identification based on multiscale representation learning"
readme = "README.md"
requires-python = ">=3.8,<3.9"
dependencies = [
    "torch==1.13.1",
    "torchvision==0.14.1",
    "pytorch-ignite==0.1.2",
    "yacs",
]

[[tool.uv.index]]
name = "pytorch-cu117"
url = "https://download.pytorch.org/whl/cu117"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu117", marker = "sys_platform == 'win32' or sys_platform == 'linux'" }
]
torchvision = [
    { index = "pytorch-cu117", marker = "sys_platform == 'win32' or sys_platform == 'linux'" }
]